refactor: 기존 CalculatorManager내 formatter 클래스를 외부 파일로 분리
